requirements.txt
alembic==1.14.0
anyio==4.7.0
email_validator==2.2.0
Faker==33.3.1
fastapi==0.115.6
httpx==0.28.1
Jinja2==3.1.5
psycopg2-binary==2.9.10
pydantic==2.10.5
python-dotenv==1.0.1
SQLAlchemy==2.0.37
starlette==0.41.3
pytest==8.3.4
bcrypt==4.2.1
passlib==1.7.4
python-jose==3.3.0